requirements.txt
fpdf2
pandas
streamlit